requirements.txt
icalendar==5.0.11
pytz==2024.1
flask==2.3.3
python-dateutil==2.8.2 